+++ changelog.md
@@ -4,6 +4,9 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 # Unreleased
 
+* Update polkadot-sdk to polkadot-stable2412-1.
+WARNING: Benchmarking command has been removed, because `frame-benchmarking-cli` crate became GPLv3 without any exception.
+
 ## Changed
 
 ## Removed